+++ Flat/config/GeneralTree.cfg
@@ -1,6 +1,7 @@
 runNumber                  int
 lumiNumber                 int
 eventNumber                uint64
+isData                     int
 npv                        int
 pu                         int
 mcWeight                   float
