+++ Flat/src/ModulesJets.cc
@@ -257,7 +257,7 @@ void PandaAnalyzer::JetBasics()
 
 // basic jet info for more jtes
 // Responsible: B. Maier
-void PandaAnalyzer::JetHbbBasics(panda::Jet& jet)
+void PandaAnalyzer::JetHbbBasics(const panda::Jet& jet)
 {
   float csv = (fabs(jet.eta())<2.5) ? jet.csv : -1;
   float cmva = (fabs(jet.eta())<2.5) ? jet.cmva : -1;
@@ -277,7 +277,7 @@ void PandaAnalyzer::JetHbbBasics(panda::Jet& jet)
 
 // stuff for b-jet energy regression
 // Responsible: B. Maier, D. Hsu
-void PandaAnalyzer::JetBRegressionInfo(panda::Jet& jet)
+void PandaAnalyzer::JetBRegressionInfo(const panda::Jet& jet)
 {
   unsigned N = cleanedJets.size()-1;
   gt->jetEMFrac[N] = jet.cef + jet.nef;
@@ -285,7 +285,7 @@ void PandaAnalyzer::JetBRegressionInfo(panda::Jet& jet)
   gt->jetLeadingLepPt[N] = 0;
   gt->jetLeadingTrkPt[N] = 0;
   gt->jetNLep[N] = 0;
-  for (const panda::Ref<panda::PFCand> &c_iter : jet.constituents) {
+  for (const panda::ConstRef<panda::PFCand> &c_iter : jet.constituents) {
     if (!c_iter.isValid())
       continue;
     auto *pf = c_iter.get();
@@ -318,7 +318,7 @@ void PandaAnalyzer::JetBRegressionInfo(panda::Jet& jet)
 
 // info for VBF jets
 // Responsible: S. Narayanan
-void PandaAnalyzer::JetVBFBasics(panda::Jet& jet)
+void PandaAnalyzer::JetVBFBasics(const panda::Jet& jet)
 {
   if (cleanedJets.size()==1) {
     jot1 = &jet;
@@ -351,7 +351,7 @@ void PandaAnalyzer::JetVBFBasics(panda::Jet& jet)
 
 // isolated jet information
 // Responsible: S. Narayanan
-void PandaAnalyzer::IsoJet(panda::Jet& jet) 
+void PandaAnalyzer::IsoJet(const panda::Jet& jet) 
 {
   float maxIsoEta = (analysis->monoh) ? 4.5 : 2.5;
   bool isIsoJet = ( (gt->nFatjet==0) || 
@@ -382,7 +382,7 @@ void PandaAnalyzer::IsoJet(panda::Jet& jet)
 
 // vary jet energy scales for jet
 // Responsible: S. Narayanan, D. Hsu
-void PandaAnalyzer::JetVaryJES(panda::Jet& jet)
+void PandaAnalyzer::JetVaryJES(const panda::Jet& jet)
 {
   // do jes variation OUTSIDE of nominal jet check 
   if (jet.ptCorrUp>jetPtThreshold) { // Vary the pT up
@@ -497,21 +497,21 @@ void PandaAnalyzer::JetVBFSystem()
 void PandaAnalyzer::JetHbbReco() 
 {
   if (centralJets.size() > 1) {
-    vector<Jet*> btagSortedJets = centralJets;
+    vector<const Jet*> btagSortedJets = centralJets;
     sort(
       btagSortedJets.begin(),
       btagSortedJets.end(),
-      analysis->useCMVA?
-        [](panda::Jet *x, panda::Jet *y) -> bool { return x->cmva > y->cmva; } :
-        [](panda::Jet *x, panda::Jet *y) -> bool { return x->csv  > y->csv ; }
+      analysis->useCMVA ?
+        [](const panda::Jet *x, const panda::Jet *y) -> bool { return x->cmva > y->cmva; } :
+        [](const panda::Jet *x, const panda::Jet *y) -> bool { return x->csv  > y->csv ; }
     );
-    map<Jet*, unsigned> order;
+    map<const Jet*, unsigned> order;
     for (unsigned i = 0; i != cleanedJets.size(); ++i) 
       order[cleanedJets[i]] = i;
 
     // the 2 best b-tagged central jets
-    panda::Jet *jet_1 = btagSortedJets.at(0);
-    panda::Jet *jet_2 = btagSortedJets.at(1);
+    const panda::Jet *jet_1 = btagSortedJets.at(0);
+    const panda::Jet *jet_2 = btagSortedJets.at(1);
     gt->hbbjtidx[0] = order[jet_1];
     gt->hbbjtidx[1] = order[jet_2];
