+++ data/deep/flatten_gen.root
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:45d5bd5fd0cb2ef4dd6e84b0bb22ebe67e76cc3700d25c9e5c953c7d908f52f7
-size 27548
+oid sha256:d3572a2022157ffbf46c6dc190714647ab3d41546a8bdd697d3d12e83377a429
+size 29967
