+++ data/deep/flatten_gen_scaled.root
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:94b46beff39664ea5a6636133c4e1955ea4757e65dadbb99bb403c877d0af992
-size 27561
+oid sha256:4819ed6dff3b83a34189e1605841fc54eab1c777cce2047b252ac2d3fdc9d83d
+size 29983
